exoscale/pyproject.toml
[build-system]
requires = [
    "hatchling>=0.13.0",
    "setuptools<61",
]
build-backend = "hatchling.build"

[project]
name = "datadog-exoscale"
description = "The Exoscale check"
readme = "README.md"
license = {text = "BSD-3-Clause"}
requires-python = ">=3.8"
keywords = [
    "datadog",
    "datadog agent",
    "datadog check",
    "exoscale",
]
authors = [
    { name = "Robert Coleman", email = "datadog@robert.net.nz" },
]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: BSD License",
    "Private :: Do Not Upload",
    "Programming Language :: Python :: 3.8",
    "Topic :: System :: Monitoring",
]
dependencies = [
    "datadog-checks-base>=25.1.0",
    "requests-exoscale-auth",
]
dynamic = [
    "version",
]

[project.optional-dependencies]
deps = []

[project.urls]
Source = "https://github.com/DataDog/integrations-extras"

[tool.hatch.version]
path = "datadog_checks/exoscale/__about__.py"

[tool.hatch.build.targets.sdist]
include = [
    "/datadog_checks",
    "/tests",
    "/manifest.json",
]

[tool.hatch.build.targets.wheel]
include = [
    "/datadog_checks/exoscale",
]
dev-mode-dirs = [
    ".",
]
